Fast_API_Project/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
PyJWT
Werkzeug
